Fix many things and finish NVT
So many small fixes, eg. the forces being calculated wrongly, the energy not coming out in eV, changed the way the distances and forces determines whether it's time to update, and many more.
Also added a function for writing Atoms objects to .xyz. Is placed in main at the moment, but will be moved.